Add semicolons to property initializers
According to the current proposal, semicolons are required. Babel 6.4 was updated to require them (see https://github.com/babel/babel/pull/3225). This is required in order to fix the build -- master will be red on Travis as soon as the build is rerun because it will pick up the latest Babel parser.