proposal for touchables as position: static
in RN proper, Touchables can be positioned absolutely or statically by positioning the children of a `Touchable`. It's as if the Touchable container serves no other purpose other than to animate the children (opacity, highlight) and provide `onPress` handlers. In RNW, all views are position `relative`, therefore children when absolutely or fix positioned move relative to the generated touchable `<button>`, rather than the view that contains the touchable component. This change will need to be applied to TouchableHightly, TouchableWithoutFeedback, etc. It will also be a breaking change for developers who utilized absolute or fixed positioning in relation to the incorrectly position relative button, but it will make RNW consistent with RN.