Update README.md
Add link and description of related project https://github.com/msand/estructor
React UI Builder, and a stack to enable in-browser cross platform development of components, pages and universal apps.
It has a component marketplace, CI/CD pipeline, and examples of most everything you need for simple projects and prototypes. It is based on Este, Structor and React Native for Web. It also integrates Apollo and CodePush into Este.