Fix modality performance issues
Fixes #1155

Inserting and deleting the modality rule triggers styles recalculation for the entire DOM tree which results in performance degradation on focus or, as the reporter suggests, makes the keyboard very slow to open in iOS.

This commit fixes the issue by toggling a scoped `rn-modality` class name on whitelisted elements or elements that get focus via keyboard and inserting a permanent rule to always disable the focus ring except for the elements that not match the class name i.e. `:focus:not(.rn-modality)`. This performs much better since the browser needs to recalculate styles only for the focuse element and its small subtree.