Image - Don't swap from defaultSource to source until fully loaded
Image provides a mechanism (defaultSource) for showing a loading or
fallback image. Currently, as soon as the source starts loading,
which is immediately after the first render, it's swapped in. The
intention is to support progressive JPEGs.

However, in cases where the user has explicitly given a defaultSource,
I'd argue that should be a better experience than watching the PJPEG load.
In conjunction with my other PR, a dev can now easily manage multiple versions
of an image via RNWs cache. Imagine on Twitter that you view a low-res image
in the timeline. Clicking to the gallery, we can read RNW's cache and provide
the highest variant found in it as the defaultSource. For users offline, or on
slow connections, the image will immediate show up and stay until we are ready
to swap in the full image. It's a roll your own PJPEG.

In addition, Enzyme now runs lifecycle methods on shallow, so I cleaned up
a few other nearby tests.