Clarify compatibility
This part of the README confused me and others--so much that some abandoned RNW and we separated. Since then, I've been making RNW apps and libraries with no problems with respect to versions. We' way beyond 0.55.X.

It only just occurred to be that 0.55.X is the minimum, not the max. Since that seems to be the case, this needs updating as it's discouraged many. Even when I had RNW apps running past `0.55`, some peers still told me it wasn't compatible. So this needs clarifying.

If indeed `0.55` is the last, um "totally compatible" version, it should still be said that it works with the latest versions, but doesn't support some features past `0.55`.