Provide better API for opening in new tab
Opening link in a new tab is something quite common on web nowadays. This use case should be handled by react-native-web.

Currently, the only solution is [described here](https://github.com/necolas/react-native-web/issues/162). And we can all agree I think that this is not what we could think as a satisfying solution.

Here is my proposal. It should be transparent for native and provide the ability for web.