Fix misspellings of NODE_ENV
I spent a good 2 hours trying to figure out why my webpack.DefinePlugin was refusing to
replace process.env.NODE_ENV in react-native-web before I realized that it had been misspelled
as NODE_DEV in a couple places. This change fixes that and adds a test for the Platform module.

This is technically a breaking change for the Platform module, although it's likely that anyone who
figured this issue out just set NODE_DEV=$NODE_ENV and will have NODE_ENV set correctly. The other
fix only controls a debug statement and is not breaking.