fix(Text): Support ellipsizeMode="clip"
The `Text` component supports different options for `ellipsizeMode`,
rather than assuming a trailing-edge '...'.

This change adds support for `ellipsizeMode="clip"` -- but only in one
special case: `numberOfLines={1}`. This is because that's all that is
supported by the browser vendors at the moment, so it's the best that
can be done by pure CSS.

Additionally, ths expands the handling of `numerOfLines` to accommodate
the value `0`, as react-native specifies that a value of `0` is [the
same as not specifying anything][rn docs].

[rn docs]: https://reactnative.dev/docs/text#numberoflines