[feat] Support title property for all elements
the `title` property is used in web for describing `<button />` without
text (for example only icons), that is used additionally to the
`aria-label`. This is because the `aria-label` is only available for the
screen readers but not for the regular user, the title will also appear
for the regular user.